REL-4170: move asterism edit util down to pot
